✨ 처음 질문 받아오는 부분 훅으로 분리
